This commit contains refactor to configuration for creation all the rabbitMQ related, all the repited code has been move to commons
