+++ management-common/src/main/java/com/ns/task/config/properties/RabbitConfig.java
@@ -1,0 +1,58 @@
+package com.ns.task.config.properties;
+
+
+import org.springframework.amqp.core.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class RabbitConfig {
+    public static final String EXCHANGE = "x.management";
+    public static final String ROUTING_KEY = "management";
+    public static final String QUEUE = "q.management.insert";
+    public static final String EXCHANGE_GET = "x.management.get";
+    public static final String ROUTING_KEY_GET = "management.get";
+    public static final String QUEUE_GET = "q.management.get";
+
+    @Bean
+    Exchange getExchange() {
+        return ExchangeBuilder.directExchange(EXCHANGE).
+                build();
+    }
+
+    @Bean
+    Binding getBinding() {
+        return BindingBuilder.bind(getQueue()).
+                to(getExchange()).
+                with(ROUTING_KEY).
+                noargs();
+    }
+
+    @Bean
+    Queue getQueue() {
+        return QueueBuilder.durable(QUEUE).
+                autoDelete().
+                build();
+    }
+
+    @Bean
+    Exchange getExchangeForGet() {
+        return ExchangeBuilder.directExchange(EXCHANGE_GET).
+                build();
+    }
+
+    @Bean
+    Binding getBindingForGet() {
+        return BindingBuilder.bind(getQueueForGet()).
+                to(getExchangeForGet()).
+                with(ROUTING_KEY_GET).
+                noargs();
+    }
+
+    @Bean
+    Queue getQueueForGet() {
+        return QueueBuilder.durable(QUEUE_GET).
+                autoDelete().
+                build();
+    }
+}
