+++ management-web/src/main/java/com/ns/task/config/Producer.java
@@ -2,6 +2,7 @@ package com.ns.task.config;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.ns.task.config.properties.RabbitConfig;
 import com.ns.task.model.ProductDTO;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -24,8 +25,8 @@ public class Producer {
 
     public void produce(ProductDTO product) throws JsonProcessingException {
         String productAsJSON = mapper.writeValueAsString(product);
-        Object o = amqpTemplate.convertSendAndReceive(ProducerConfig.EXCHANGE, ProducerConfig.ROUTING_KEY, productAsJSON);
+        Object o = amqpTemplate.convertSendAndReceive(RabbitConfig.EXCHANGE, RabbitConfig.ROUTING_KEY, productAsJSON);
         logger.info("Sending message {}", o);
-        logger.info("Sending message = {} to queue {}", productAsJSON, ProducerConfig.QUEUE);
+        logger.info("Sending message = {} to queue {}", productAsJSON, RabbitConfig.QUEUE);
     }
 }
